refactor: restructure folder by feature
