+++ README.md
@@ -1,0 +1,13 @@
+# Bis Develop Tool
+
+## 環境需求
+- Node.js v18+
+- npm
+
+## 安裝
+```bash
+cd .\frontend
+npm install
+
+cd .\backend
+npm run dev
