+++ backend/src/services/parserService.ts
