+++ backend/src/services/pathService.ts
