+++ frontend/src/App.tsx
@@ -1,6 +1,6 @@
 import React from 'react';
 import { BrowserRouter as Router, Route, Routes } from 'react-router-dom';
-import routes from './routes';
+import routes from './routes/routes';
 
 // function App() {
 //   return <Home />;
