+++ frontend/src/features/compressed/services/compressedService.ts
