+++ frontend/src/features/compressed/types.ts
@@ -1,0 +1,6 @@
+export interface Files {
+  files: string[];
+  zipPath: string;
+  zipName: string;
+  delList: string[];
+}
