+++ frontend/src/features/hostMsg/utils/xmlParser.ts
