+++ frontend/src/features/test/types.ts
@@ -1,0 +1,4 @@
+export interface User {
+  id: number;
+  name: string;
+}
