+++ frontend/src/types/types.ts
@@ -1,0 +1,4 @@
+export interface Path {
+  path: string;
+  name: string;
+}
